requirement.txt
openai
pillow
loguru
python-dotenv
opencv-python
gradio